requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
anyio==3.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.2
click==8.1.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
discord.py==1.7.3
Django==4.0.3
djangorestframework==3.13.1
entrypoints==0.3
fastapi==0.73.0
Flask==2.1.1
future==0.18.2
h11==0.13.0
idna==3.2
importlib-metadata==4.11.3
imutils==0.5.4
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
iso8601==1.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
multidict==5.1.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
neo4j==4.3.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
pika==1.2.0
Pillow==8.3.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyenchant==3.2.2
PyGetWindow==0.0.9
Pygments==2.11.2
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.18.0
pytweening==1.0.4
pytz==2021.1
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
redis==4.2.1
requests==2.26.0
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.0
tzdata==2022.1
urllib3==1.26.6
uvicorn==0.17.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.1.1
wrapt==1.14.0
yarl==1.6.3
zipp==3.7.0